extract internal tests (+ typo)
