+++ widget/entry_internal_test.go
@@ -1,0 +1,130 @@
+package widget
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+
+	"fyne.io/fyne"
+	"fyne.io/fyne/driver/desktop"
+	"fyne.io/fyne/test"
+	"fyne.io/fyne/theme"
+)
+
+func TestEntry_ExpandSelectionForDoubleTap(t *testing.T) {
+	str := []rune(" fish 日本語日  \t  test 本日本 moose  \t")
+
+	// select invalid (before start)
+	start, end := getTextWhitespaceRegion(str, -1)
+	assert.Equal(t, -1, start)
+	assert.Equal(t, -1, end)
+
+	// select whitespace at the end of text
+	start, end = getTextWhitespaceRegion(str, len(str))
+	assert.Equal(t, 29, start)
+	assert.Equal(t, 32, end)
+	start, end = getTextWhitespaceRegion(str, len(str)+100)
+	assert.Equal(t, 29, start)
+	assert.Equal(t, 32, end)
+
+	// select the whitespace
+	start, end = getTextWhitespaceRegion(str, 0)
+	assert.Equal(t, 0, start)
+	assert.Equal(t, 1, end)
+
+	// select "fish"
+	start, end = getTextWhitespaceRegion(str, 1)
+	assert.Equal(t, 1, start)
+	assert.Equal(t, 5, end)
+	start, end = getTextWhitespaceRegion(str, 4)
+	assert.Equal(t, 1, start)
+	assert.Equal(t, 5, end)
+
+	// select "日本語日"
+	start, end = getTextWhitespaceRegion(str, 6)
+	assert.Equal(t, 6, start)
+	assert.Equal(t, 10, end)
+	start, end = getTextWhitespaceRegion(str, 9)
+	assert.Equal(t, 6, start)
+	assert.Equal(t, 10, end)
+
+	// select "  \t  "
+	start, end = getTextWhitespaceRegion(str, 10)
+	assert.Equal(t, 10, start)
+	assert.Equal(t, 15, end)
+
+	// select "  \t"
+	start, end = getTextWhitespaceRegion(str, 30)
+	assert.Equal(t, 29, start)
+	assert.Equal(t, len(str), end)
+}
+
+func TestEntry_ExpandSelectionWithWordSeparators(t *testing.T) {
+	// select "is_a"
+	str := []rune("This-is_a-test")
+	start, end := getTextWhitespaceRegion(str, 6)
+	assert.Equal(t, 5, start)
+	assert.Equal(t, 9, end)
+}
+
+func TestEntry_EraseSelection(t *testing.T) {
+	// Selects "sti" on line 2 of a new multiline
+	// T e s t i n g
+	// T e[s t i]n g
+	// T e s t i n g
+	e := NewMultiLineEntry()
+	e.SetText("Testing\nTesting\nTesting")
+	e.CursorRow = 1
+	e.CursorColumn = 2
+	var keyDown = func(key *fyne.KeyEvent) {
+		e.KeyDown(key)
+		e.TypedKey(key)
+	}
+	var keyPress = func(key *fyne.KeyEvent) {
+		keyDown(key)
+		e.KeyUp(key)
+	}
+	keyDown(&fyne.KeyEvent{Name: desktop.KeyShiftLeft})
+	keyPress(&fyne.KeyEvent{Name: fyne.KeyRight})
+	keyPress(&fyne.KeyEvent{Name: fyne.KeyRight})
+	keyPress(&fyne.KeyEvent{Name: fyne.KeyRight})
+
+	e.eraseSelection()
+	e.updateText(e.textProvider().String())
+	assert.Equal(t, "Testing\nTeng\nTesting", e.Text)
+	a, b := e.selection()
+	assert.Equal(t, -1, a)
+	assert.Equal(t, -1, b)
+}
+
+func TestEntry_MouseClickAndDragOutsideText(t *testing.T) {
+	entry := NewEntry()
+	entry.SetText("A\nB\n")
+
+	testCharSize := theme.TextSize()
+	pos := fyne.NewPos(testCharSize, testCharSize*4) // tap below rows
+	ev := &fyne.PointEvent{Position: pos}
+
+	me := &desktop.MouseEvent{PointEvent: *ev, Button: desktop.LeftMouseButton}
+	entry.MouseDown(me)
+	de := &fyne.DragEvent{PointEvent: *ev, DraggedX: 1, DraggedY: 0}
+	entry.Dragged(de)
+	entry.MouseUp(me)
+	assert.False(t, entry.selecting)
+}
+
+func TestEntry_PasteFromClipboard(t *testing.T) {
+	entry := NewEntry()
+
+	w := test.NewApp().NewWindow("")
+	w.SetContent(entry)
+
+	testContent := "test"
+
+	clipboard := fyne.CurrentApp().Driver().AllWindows()[0].Clipboard()
+	clipboard.SetContent(testContent)
+
+	entry.pasteFromClipboard(clipboard)
+
+	assert.Equal(t, entry.Text, testContent)
+}
